requirements.txt
tensorflow>=2.10.0
flask>=2.3.0
werkzeug>=2.3.0
opencv-python>=4.7.0
numpy>=1.23.0
Pillow>=9.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
